pyproject.toml
[tool.poetry]
name = "vraagslaaf"
version = "0.1.0"
description = ""
authors = ["Anton Leagre <anton.leagre@hotmail.com>"]
packages = [
    { include = "Vraagslaaf" },
]

[tool.poetry.dependencies]
python = "^3.9"
discord = "^1.0.1"
discord_components = "^1.0.0"
flask = "^1.1.2"
PyNaCl = "^1.4.0"
PyMuPDF = "^1.18.5"
requests = "^2.25.0"
google-auth-oauthlib = "^0.4.2"
google-api-python-client = "^1.12.8"
quart = "^0.14.1"
python-dotenv = "^0.18.0"
gunicorn = "^20.1.0"
websockets = "^9.1"
Hypercorn = "^0.11.2"
discord-py-slash-command = "^2.4.0"
oauth2client = "3.0.0"
google-auth-httplib2 = "^0.1.0"
aiogoogle = "^2.0.1"
ffmpeg = "^1.4"
youtube_dl = "^2021.6.6"
feedparser = "^6.0.8"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
bot = "Vraagslaaf.main:main"
server = "Vraagslaaf.server:runner"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
